# faithgpt-platform/api/openapi.yaml
openapi: 3.0.3
info:
  title: FaithGPT API
  version: "1.0.0"
  description: >
    Canonical REST API contract for FaithGPT™ ("Transforming Scripture into Daily Living").
    Conventions, envelopes, pagination, streaming, and rate-limiting are documented narratively
    in /faithgpt-platform/docs/09-api-architecture.md — this document is the reviewable
    machine-readable snapshot. Schema names mirror /faithgpt-platform/schema/schema.prisma
    verbatim (see /faithgpt-platform/docs/00-shared-reference.md §13).
  contact:
    name: FaithGPT Engineering
    url: https://faithgpt.app
servers:
  - url: https://api.faithgpt.app/api/v1
    description: Production
  - url: https://staging-api.faithgpt.app/api/v1
    description: Staging
  - url: http://localhost:3000/api/v1
    description: Local development
...
security:
  - bearerAuth: []

# =====================================================================================
# PATHS
# =====================================================================================
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user (email/password)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, name]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 10 }
                name: { type: string }
                denominationLens: { type: string, nullable: true, example: "NEUTRAL" }
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        "409":
          $ref: '#/components/responses/Conflict'
        "400":
          $ref: '#/components/responses/ValidationError'
  /auth/login:
    post:
      tags: [Auth]
      summary: Login with email/password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
                deviceInfo: { type: string, nullable: true }
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        "401":
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Exchange a refresh token for a new access/refresh token pair (rotation)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        "200":
          description: New token pair issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        "401":
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      tags: [Auth]
      summary: Revoke a refresh token (log out this device)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        "204":
          description: Logged out
  /auth/oauth/{provider}/callback:
    post:
      tags: [Auth]
      summary: Exchange a Google/Apple/Facebook identity token for a FaithGPT session
      security: []
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [google, apple, facebook]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [idToken]
              properties:
                idToken: { type: string }
                deviceInfo: { type: string, nullable: true }
      responses:
        "200":
          description: Authenticated (existing account linked or new account created)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        "401":
          $ref: '#/components/responses/Unauthorized'
  /auth/verify-phone:
    post:
      tags: [Auth]
      summary: Verify a phone number via Twilio Verify OTP, or sign in/up with PHONE_OTP
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, code]
              properties:
                phone: { type: string, example: "+233201234567" }
                code: { type: string, example: "123456" }
                deviceInfo: { type: string, nullable: true }
      responses:
        "200":
          description: Verified / authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        "401":
          $ref: '#/components/responses/Unauthorized'
  /auth/verify-email:
    get:
      tags: [Auth]
      summary: Verify email via single-use token (24h expiry)
      security: []
      parameters:
        - name: token
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Email verified
        "401":
          $ref: '#/components/responses/Unauthorized'
  /auth/password-reset/request:
    post:
      tags: [Auth]
      summary: Request a password reset email (always returns 202)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        "202":
          description: Accepted (no account-existence leakage)
  /auth/password-reset/confirm:
    post:
      tags: [Auth]
      summary: Confirm a password reset, revoking all existing refresh tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, newPassword]
              properties:
                token: { type: string }
                newPassword: { type: string, minLength: 10 }
      responses:
        "200":
          description: Password updated, all sessions revoked
        "401":
          $ref: '#/components/responses/Unauthorized'

  # ---------------------------------------------------------------------------------
  # USERS
  # ---------------------------------------------------------------------------------
...
  /bible/books/{bookId}/chapters/{chapter}:
    get:
      tags: [Bible]
      summary: Get all verses for a chapter in a given version
      security: []
      parameters:
        - name: bookId
          in: path
          required: true
          schema: { type: string }
        - name: chapter
          in: path
          required: true
          schema: { type: integer }
        - name: version
          in: query
          required: false
          schema: { type: string, default: "KJV", example: "ESV" }
      responses:
        "200":
          description: Verses with text for the requested version
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/VerseWithText'
        "404":
          $ref: '#/components/responses/NotFound'
  /bible/search:
    get:
      tags: [Bible]
      summary: Full-text search across Bible text (by version)
      security: []
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string }
        - name: version
          in: query
          required: false
          schema: { type: string, default: "KJV" }
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        "200":
          description: Matching verses
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/VerseWithText'
                  meta:
                    $ref: '#/components/schemas/CursorPaginationMeta'

  # ---------------------------------------------------------------------------------
  # BIBLE INTERACTIONS (Highlights / Notes / Bookmarks)
  # ---------------------------------------------------------------------------------
  /highlights:
    get:
      tags: [Bible Interactions]
      summary: List the authenticated user's highlights
      parameters:
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        "200":
          description: Highlights
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Highlight'
                  meta:
                    $ref: '#/components/schemas/CursorPaginationMeta'
    post:
      tags: [Bible Interactions]
      summary: Highlight a verse
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [verseId]
              properties:
                verseId: { type: string }
                color: { type: string, default: "yellow" }
      responses:
        "201":
          description: Highlight created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Highlight'
  /highlights/{id}:
    delete:
      tags: [Bible Interactions]
      summary: Remove a highlight
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        "204":
          description: Removed
...
  /bookmarks:
    get:
      tags: [Bible Interactions]
      summary: List the authenticated user's bookmarks
      parameters:
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        "200":
          description: Bookmarks
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Bookmark'
                  meta:
                    $ref: '#/components/schemas/CursorPaginationMeta'
    post:
      tags: [Bible Interactions]
      summary: Bookmark a verse
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [verseId]
              properties:
                verseId: { type: string }
      responses:
        "201":
          description: Bookmark created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Bookmark'
  /bookmarks/{id}:
    delete:
      tags: [Bible Interactions]
      summary: Remove a bookmark
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        "204":
          description: Removed

  # ---------------------------------------------------------------------------------
  # READING PLANS
  # ---------------------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema: { type: string }
      description: Opaque cuid()
    CursorParam:
      name: cursor
      in: query
      required: false
      schema: { type: string }
    LimitParam:
      name: limit
      in: query
      required: false
      schema: { type: integer, default: 20, maximum: 100 }
...
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthorized:
      description: Missing/invalid credentials or token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Conflict:
      description: Resource conflict (e.g., unique constraint)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
...
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: "VALIDATION_ERROR" }
            message: { type: string }
            details:
              type: object
              nullable: true
              additionalProperties: true
    CursorPaginationMeta:
      type: object
      properties:
        pagination:
          type: object
          properties:
            nextCursor: { type: string, nullable: true }
            hasMore: { type: boolean }
...
    AuthTokenResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            accessToken: { type: string }
            refreshToken: { type: string }
            user:
              $ref: '#/components/schemas/User'
...
    UserRole:
      type: string
      enum: [USER, GROUP_LEADER, MINISTRY_ADMIN, MODERATOR, CONTENT_EDITOR, SUPPORT, ADMIN, SUPER_ADMIN]
    SubscriptionTier:
      type: string
      enum: [FREE, PLUS, PREMIUM, MINISTRY, LIFETIME]
...
    User:
      type: object
      properties:
        id: { type: string }
        email: { type: string, format: email, nullable: true }
        phone: { type: string, nullable: true }
        name: { type: string }
        avatarUrl: { type: string, nullable: true }
        role:
          $ref: '#/components/schemas/UserRole'
        tier:
          $ref: '#/components/schemas/SubscriptionTier'
        denominationLens: { type: string, nullable: true, example: "NEUTRAL" }
        locale: { type: string, example: "en" }
        timezone: { type: string, example: "UTC" }
        defaultBibleVersionId: { type: string, nullable: true }
        emailVerifiedAt: { type: boolean }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
...
    VerseWithText:
      type: object
      properties:
        id: { type: string }
        bookId: { type: string }
        chapter: { type: integer }
        verse: { type: integer }
        reference: { type: string, example: "Romans 8:28" }
        text: { type: string, example: "And we know that for those who love God all things work together for good..." }
        versionCode: { type: string, example: "ESV" }
    Highlight:
      type: object
      properties:
        id: { type: string }
        verseId: { type: string }
        color: { type: string, example: "yellow" }
        createdAt: { type: string, format: date-time }
...
    Bookmark:
      type: object
      properties:
        id: { type: string }
        verseId: { type: string }
        createdAt: { type: string, format: date-time }

    # --- Reading Plans ---------------------------------------------------------
...